Move pmts and arapucas classes in separate file
